services/engine-grounding/pyproject.toml
[project]
name = "engine-grounding"
version = "0.1.0"
description = "ZROKY Grounding Engine — RAG quality, hallucination detection, claim attribution"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic-settings>=2.7.1",
    "google-cloud-pubsub>=2.29.0",
    "redis>=5.2.0",
    "clickhouse-connect>=0.8.0",
    "psycopg[binary]>=3.2.0",
    "httpx>=0.28.1",
    "numpy>=1.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.0",
    "httpx>=0.28.1",
    "ruff>=0.9.4",
    "mypy>=1.14.1",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
